@@ -1,6 +1,7 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
+using System.Text.Json;
 using System.Text.Json.Nodes;
 
 namespace FunctionsNetHost.Grpc
@@ -12,30 +13,41 @@ namespace FunctionsNetHost.Grpc
         /// Builds the path by reading the worker.config.json
         /// </summary>
         /// <param name="applicationDirectory">The FunctionAppDirectory value from environment reload request.</param>
-        internal static string GetApplicationExePath(string applicationDirectory)
+        internal static string? GetApplicationExePath(string applicationDirectory)
         {
-            var workerConfigPath = Path.Combine(applicationDirectory, "worker.config.json");
-
-            if (!File.Exists(workerConfigPath))
+            string jsonString = string.Empty;
+            string workerConfigPath = string.Empty;
+            try
             {
-                throw new FileNotFoundException($"worker.config.json file not found", fileName: workerConfigPath);
-            }
+                workerConfigPath = Path.Combine(applicationDirectory, "worker.config.json");
 
-            if (Logger.IsDebugLogEnabled)
-            {
-                Logger.LogDebug($"workerConfigPath:{workerConfigPath}");
-            }
+                jsonString = File.ReadAllText(workerConfigPath);
+                var workerConfigJsonNode = JsonNode.Parse(jsonString)!;
+                var executableName = workerConfigJsonNode["description"]?["defaultWorkerPath"]?.ToString();
 
-            var jsonString = File.ReadAllText(workerConfigPath);
-            var workerConfigJsonNode = JsonNode.Parse(jsonString)!;
-            var executableName = workerConfigJsonNode["description"]?["defaultWorkerPath"]?.ToString();
-            
-            if (executableName == null)
-            {
-                throw new InvalidOperationException("Invalid worker configuration.");
+                if (executableName == null)
+                {
+                    Logger.Log($"Invalid worker configuration. description > defaultWorkerPath property value is null. jsonString:{jsonString}");
+                    return null;
+                }
+
+                return Path.Combine(applicationDirectory, executableName);
             }
-            
-            return Path.Combine(applicationDirectory, executableName);
+            catch (FileNotFoundException ex)
+            {
+                Logger.Log($"{workerConfigPath} file not found.{ex}");
+                return null;
+            }
+            catch (JsonException ex)
+            {
+                Logger.Log($"Error parsing JSON in GetApplicationExePath.{ex}. jsonString:{jsonString}");
+                return null;
+            }
+            catch (Exception ex)
+            {
+                Logger.Log($"Error in GetApplicationExePath.{ex}. jsonString:{jsonString}");
+                return null;
+            }
         }
     }
 }
