@@ -5,43 +5,44 @@ using System.Runtime.InteropServices;
 
 namespace FunctionsNetHost
 {
+    /// <summary>
+    /// Manages loading hostfxr & worker assembly.
+    /// </summary>
     internal sealed class AppLoader : IDisposable
     {
-        private static readonly AppLoader _instance = new();
         private IntPtr _hostfxrHandle = IntPtr.Zero;
+        private IntPtr _hostContextHandle = IntPtr.Zero;
         private bool _disposed;
 
-        private AppLoader()
+        internal AppLoader()
         {
             LoadHostfxrLibrary();
         }
 
-        internal static AppLoader Instance => _instance;
-
         private void LoadHostfxrLibrary()
         {
             // If having problems with the managed host, enable the following:
-            //Environment.SetEnvironmentVariable("COREHOST_TRACE", "1");
+            // Environment.SetEnvironmentVariable("COREHOST_TRACE", "1");
             // In Unix environment, you need to run the below command in the terminal to set the environment variable.
             // export COREHOST_TRACE=1
 
-            var hostfxrFullPath = PathResolver.GetHostFxrPath();
-            Logger.LogDebug($"hostfxrFullPath:{hostfxrFullPath}");
+            var hostfxrFullPath = NetHost.GetHostFxrPath();
+            Logger.LogTrace($"hostfxr path:{hostfxrFullPath}");
 
             _hostfxrHandle = NativeLibrary.Load(hostfxrFullPath);
             if (_hostfxrHandle == IntPtr.Zero)
             {
-                Logger.LogInfo($"Failed to load hostfxr. hostfxrFullPath:{hostfxrFullPath}");
+                Logger.Log($"Failed to load hostfxr. hostfxr path:{hostfxrFullPath}");
                 return;
             }
 
-            Logger.LogDebug($"hostfxr library loaded successfully.");
+            Logger.LogTrace($"hostfxr library loaded successfully.");
         }
 
-        internal int RunApplication(string assemblyPath)
+        internal int RunApplication(string? assemblyPath)
         {
-            Logger.LogDebug($"Assembly path to run:{assemblyPath}");
-            
+            ArgumentNullException.ThrowIfNull(assemblyPath, nameof(assemblyPath));
+
             unsafe
             {
                 var parameters = new HostFxr.hostfxr_initialize_parameters
@@ -49,12 +50,12 @@ namespace FunctionsNetHost
                     size = sizeof(HostFxr.hostfxr_initialize_parameters)
                 };
 
-                var error = HostFxr.Initialize(1, new[] { assemblyPath }, ref parameters, out var hostContextHandle);
+                var error = HostFxr.Initialize(1, new[] { assemblyPath }, ref parameters, out _hostContextHandle);
 
-                if (hostContextHandle == IntPtr.Zero)
+                if (_hostContextHandle == IntPtr.Zero)
                 {
-                    Logger.LogInfo(
-                        $"Failed to initialize the .NET Core runtime. assemblyPath:{assemblyPath}");
+                    Logger.Log(
+                        $"Failed to initialize the .NET Core runtime. Assembly path:{assemblyPath}");
                     return -1;
                 }
 
@@ -63,9 +64,10 @@ namespace FunctionsNetHost
                     return error;
                 }
 
-                HostFxr.SetAppContextData(hostContextHandle, "AZURE_FUNCTIONS_NATIVE_HOST", "1");
+                Logger.LogTrace($"hostfxr initialized with {assemblyPath}");
+                HostFxr.SetAppContextData(_hostContextHandle, "AZURE_FUNCTIONS_NATIVE_HOST", "1");
 
-                return HostFxr.Run(hostContextHandle);
+                return HostFxr.Run(_hostContextHandle);
             }
         }
 
@@ -87,8 +89,15 @@ namespace FunctionsNetHost
                 if (_hostfxrHandle != IntPtr.Zero)
                 {
                     NativeLibrary.Free(_hostfxrHandle);
-                    Logger.LogInfo($"Freed hostfxr library handle");
+                    Logger.LogTrace($"Freed hostfxr library handle");
                     _hostfxrHandle = IntPtr.Zero;
+                }
+
+                if (_hostContextHandle != IntPtr.Zero)
+                {
+                    NativeLibrary.Free(_hostContextHandle);
+                    Logger.LogTrace($"Freed hostcontext handle");
+                    _hostContextHandle = IntPtr.Zero;
                 }
 
                 _disposed = true;
