@@ -5,18 +5,33 @@ namespace FunctionsNetHost
 {
     internal static class EnvironmentUtils
     {
+#if OS_LINUX
+        [System.Runtime.InteropServices.DllImport("libc")]
+        private static extern int setenv(string name, string value, int overwrite);
+#endif
+
         /// <summary>
         /// Gets the environment variable value.
         /// </summary>
         internal static string? GetValue(string environmentVariableName)
         {
-            var value = Environment.GetEnvironmentVariable(environmentVariableName);
-            if (Logger.IsDebugLogEnabled)
-            {
-                Logger.LogDebug($"{environmentVariableName} environment variable value:{value}");
-            }
+            return Environment.GetEnvironmentVariable(environmentVariableName);
+        }
 
-            return value;
+        /// <summary>
+        /// Sets the environment variable value.
+        /// </summary>
+        internal static void SetValue(string name, string value)
+        {
+            /*
+             *  Environment.SetEnvironmentVariable is not setting the value of the parent process in Unix.
+             *  So using the native method directly here.
+             * */
+#if OS_LINUX
+            setenv(name, value, 1);
+#else
+            Environment.SetEnvironmentVariable(name, value);
+#endif
         }
     }
 }
