@@ -12,8 +12,8 @@ namespace FunctionsNetHost
         [UnmanagedCallersOnly(EntryPoint = "get_application_properties")]
         public static int GetApplicationProperties(NativeHostData nativeHostData)
         {
-            Logger.LogDebug("NativeExports.GetApplicationProperties method invoked.");
-            
+            Logger.LogTrace("NativeExports.GetApplicationProperties method invoked.");
+
             try
             {
                 var nativeHostApplication = NativeHostApplication.Instance;
@@ -25,8 +25,8 @@ namespace FunctionsNetHost
             }
             catch (Exception ex)
             {
-                Logger.LogInfo($"Error in NativeExports.GetApplicationProperties: {ex}");
-                return 0; 
+                Logger.Log($"Error in NativeExports.GetApplicationProperties: {ex}");
+                return 0;
             }
         }
 
@@ -35,8 +35,8 @@ namespace FunctionsNetHost
                                                 delegate* unmanaged<byte**, int, IntPtr, IntPtr> requestCallback,
             IntPtr grpcHandler)
         {
-            Logger.LogDebug("NativeExports.RegisterCallbacks method invoked.");
-            
+            Logger.LogTrace("NativeExports.RegisterCallbacks method invoked.");
+
             try
             {
                 NativeHostApplication.Instance.SetCallbackHandles(requestCallback, grpcHandler);
@@ -44,7 +44,7 @@ namespace FunctionsNetHost
             }
             catch (Exception ex)
             {
-                Logger.LogInfo($"Error in RegisterCallbacks: {ex}");
+                Logger.Log($"Error in RegisterCallbacks: {ex}");
                 return 0;
             }
         }
@@ -52,8 +52,6 @@ namespace FunctionsNetHost
         [UnmanagedCallersOnly(EntryPoint = "send_streaming_message")]
         public static unsafe int SendStreamingMessage(IntPtr pInProcessApplication, byte* streamingMessage, int streamingMessageSize)
         {
-            Logger.LogDebug($"NativeExports.SendStreamingMessage method invoked.");
-
             try
             {
                 var span = new ReadOnlySpan<byte>(streamingMessage, streamingMessageSize);
@@ -65,7 +63,7 @@ namespace FunctionsNetHost
             }
             catch (Exception ex)
             {
-                Logger.LogInfo($"Error in SendStreamingMessage: {ex}");
+                Logger.Log($"Error in SendStreamingMessage: {ex}");
                 return 0;
             }
         }
